Add support to more functions, special numbers and enriched README file:
Added support for pi and e numbers. The implementation works well but feels like kind of a hack.
In the future I would like to re-design the parser to handle these cases in a more elegant way.

Added support for logarithms with base different than the e number.
The syntaxis is "\log_x(y)" where x is the desired base and y the argument.

Improved README file. Examples are clearer and included notes about allowed syntax.