Fix important bug
There was an important bug, one deep in the Parser.
It turns out that I was trying to use FRAC as
a binary expression and that was causing problems
when the input was alike "\frac{x}{y}*z"; the parser just returned the fraction.

Now the FRAC token is handled in the bottom of the
parser (more like a grouping expression).